Move transformer methods to `shiny.render.transformer` folder
Also move `RenderFunction` methods into `shiny.render._deprecated`
